second commit

+++ lib/functions/db_functions/db_functions.dart
@@ -1,10 +1,51 @@
+import 'package:flutter/material.dart';
 import 'package:hive/hive.dart';
 import '../../Model/video_model/video_model.dart';
 
-//video_model db function-----------
+//delete from db with a confirmation show dailog..
 
-//delete from db
-Future<void> deletefromDB(int id) async {
+Future<void> deleteFromDB(BuildContext context, int id) async {
   final videoDB = await Hive.openBox<VideoModel>('videos');
-  await videoDB.deleteAt(id);
+
+  // ignore: use_build_context_synchronously
+  bool confirmDelete = await showDialog(
+    context: context,
+    builder: (BuildContext context) {
+      return AlertDialog(
+        backgroundColor: Colors.black,
+        title: const Text(
+          "Confirm Deletion",
+          style: TextStyle(color: Colors.white),
+        ),
+        content: const Text(
+          "Are you sure you want to delete this video?",
+          style: TextStyle(color: Colors.white),
+        ),
+        actions: <Widget>[
+          TextButton(
+            child: Text(
+              "No",
+              style: TextStyle(color: Colors.orange.shade700),
+            ),
+            onPressed: () {
+              Navigator.of(context).pop(false);
+            },
+          ),
+          TextButton(
+            child: Text(
+              "Yes",
+              style: TextStyle(color: Colors.orange.shade700),
+            ),
+            onPressed: () {
+              Navigator.of(context).pop(true);
+            },
+          ),
+        ],
+      );
+    },
+  );
+
+  if (confirmDelete == true) {
+    await videoDB.deleteAt(id);
+  }
 }
